- refactor data layer to fetch all images and videos then map them into albums - add unit test for repository

+++ app/src/main/java/com/apdallahyousry/customgalleryapplication/data/models/MediaItemModel.kt
@@ -1,6 +1,6 @@
 package com.apdallahyousry.customgalleryapplication.data.models
 
-data class MediaItemModel(val id:Int,
+data class MediaItemModel(val id:Long,
                           val title:String,
                           val albumName:String?=null,
                           val mediaPath:String?=null,
